Assert an execution unit has not more that 'u16::MAX' children

+++ src/intermediate_tree.rs
@@ -234,7 +234,7 @@ pub enum NodeVariant {
     ReadSymbol(Identifier),
 
     // --- Children function access
-    ChildFunRef(usize),
+    ChildFunRef(u16),
 
     // --- Literals
     NullLiteral,
